backend/requirements.txt
# ===== Django Core =====
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0

# ===== Deployment =====
gunicorn==23.0.0          # WSGI server for production
whitenoise==6.11.0        # Static file serving
dj-database-url==3.0.1    # DB URL parsing
psycopg2-binary==2.9.11   # PostgreSQL driver

# ===== Utilities =====
requests==2.32.5           # HTTP requests
python-dateutil==2.9.0.post0  # Date utilities
pytz==2025.2               # Timezone support
sqlparse==0.5.3            # SQL formatting (used by Django)
